% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addROIfromKey.R
\name{addROIfromKey}
\alias{addROIfromKey}
\title{Add ROIs generated from RegToolboxMSRC's IMS processing}
\usage{
addROIfromKey(cardinaldata, ROI_key, overlap_thresh = 0.5, xy_min = F,
  show_images = F)
}
\arguments{
\item{cardinaldata}{MSImageSet to which ROIs are added}

\item{ROI_key}{data.frame imported with importSpatialCorrKey}

\item{overlap_thresh}{numeric overlap percentage threshold between 0 and 1, set to 0 to load in all}

\item{xy_min}{logical whether the MSImageSet coordinates are reduced (i.e. min x and y coordinates are equal to 1) or in their original configuration}

\item{show_images}{logical plot binary masks of each imported ROI (takes longer to process)}
}
\value{
MSImageSet with ROIs are added
}
\description{
Adds ROIs to MSImageSet, ROIs are added in the pixelData of the MSImageSet as three vectors. First, the ROI is combined with it's index and name to make a unique identified that is added a logical vector where TRUE pixels are in the ROI regardless of degree of overlap. Another vector is added for the ROI per-pixel overlap where values of 0 indicate no overlap, and values of 1 indicate full overlap. Finally, the integer index of the ROI can be accessed by the '_roi_index' appended vector.
}
