% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genUniqueXYcardinal.R
\name{genUniqueXYcardinal}
\alias{genUniqueXYcardinal}
\title{Return unique xy identifier}
\usage{
genUniqueXYcardinal(cardinaldata)
}
\arguments{
\item{cardinaldata}{MSImageSet for data analysis}
}
\value{
character vector with unique XY identifier for matching
}
\description{
Return unique xy identifier
}
