% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSpatialCorrKey.R
\name{importSpatialCorrKey}
\alias{importSpatialCorrKey}
\title{Import Spatial Correlation IMS key}
\usage{
importSpatialCorrKey(filepath, overlap_thresh = 0.5)
}
\arguments{
\item{filepath}{path to the spatial correlation IMS key .csv}

\item{overlap_thresh}{overlap percentage threshold between 0 and 1, set to 0 to load in all}
}
\value{
data.frame of thresholded spatial correlation key
}
\description{
Reads spatial correlation IMS key as data.frame and thresholds pixels below a certain overlap percentage
}
