% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceXYcardinal.R
\name{reduceXYcardinal}
\alias{reduceXYcardinal}
\title{Reduce x,y coordinate offsets for Cardinal MSImageSet}
\usage{
reduceXYcardinal(cardinaldata)
}
\arguments{
\item{cardinaldata}{a Cardinal MSImageSet of full profile data}
}
\value{
x,y reduced MSImageSet
}
\description{
sets minimum cardinal x,y coordinates to 1
}
